[plumb.el] Now use url-retrieve instead of start-process.
Added variables `plumb-image-directory' and `plumb-pdf-directory'.
Images and pdfs are downloaded in their respective locations instead
of /tmp.

Added new option in `plumb': "Download and open", which simply fetches
the url and opens it in a new buffer.  Useful for links that you know
are plaintext so you don't need to use eww.