[eshell] Removed clear-please and delete-or-quit functions.
I think I've found a better workflow (hype word of the century) in
eshell-outline-mode.

Rather than killing eshell buffers left and right, it is more useful
to keep one or two large eshell buffers and use narrowing and outlines
to keep it under control.

As such, I recently made C-c M-o narrow to the current input+output.
This is in my opinion a far more useful alternative to inserting a
bunch of newlines.